Add files via upload
Ok, so you have built your 37th space station and keep needing larger stations. You have built monstrous craft to try and find a way to lift most of the station in one go. By the time the craft is done, it contains hundreds of parts. You launch it, and your system starts laughing uncontrollably,  and finally manages to say “You want to go where?” That is without any payload yet.

If that describes you, then this is for you. You have proved you can do what it takes, you now just want to move on and colonize the universe.

This is a single part SSTO cargo hauler. It started out as a Delta Wing, I swear. Don't even ask how it ended up this shape. Any way. It attempts to be all things to all men at all times. It contains most of what you need to get to orbit. I was tired of long thin transport vehicles, so it has a large squarish cargo bay. Still on the flat side, the craft does have to be at least wingish in shape. The top opens up either as 2 gull wings, or as a whole. Some form of